% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRApp-package.R
\docType{package}
\name{FRApp-package}
\alias{FRApp-package}
\title{FRApp: FRAP Data Analysis Using Nonlinear Mixed Effect Models with Shiny}
\description{
Analysis of Fluorescence Recovery After Photobleaching (FRAP) experiments using nonlinear mixed-effects regression models and analysis of the results. FRApp is not limited to the analysis of FRAP experiments only. Any nonlinear mixed-effects models with an asymptotic exponential functional relationship to hierarchical data in various domains can be fitted. The analysis of data available in the package is presented in Di Credico, G., Pelucchi, S., Pauli, F. et al. (2025) \doi{10.1038/s41598-025-87154-w}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gioiadc/FRApp}
  \item Report bugs at \url{https://github.com/gioiadc/FRApp/issues}
}

}
\author{
\strong{Maintainer}: Gioia Di Credico \email{gioia.dicredico@deams.units.it} (\href{https://orcid.org/0000-0001-7502-4160}{ORCID}) [copyright holder]

}
\keyword{internal}
