% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRApp.R
\name{FRApp}
\alias{FRApp}
\title{FRApp}
\usage{
FRApp(...)
}
\arguments{
\item{...}{shinyApp function options}
}
\value{
A list of objects that defines the Shiny app.
}
\description{
A Shiny app to fit the nonlinear mixed effects model to FRAP data. By default,
R opens the Shiny app in the default browser.
}
\examples{
# Run the app
\dontrun{
FRApp()
}
}
